+++ src/holodeck/command.py
@@ -6,6 +6,7 @@ worry about these.
 
 
 import numpy as np
+import sys
 from holodeck.exceptions import HolodeckException
 
 
@@ -210,13 +211,22 @@ class SpawnAgentCommand(Command):
 
     """
 
-    def __init__(self, location, rotation, name, agent_type, is_main_agent=False):
+    def __init__(
+        self,
+        location,
+        rotation,
+        name,
+        agent_type,
+        max_height=sys.maxsize,
+        is_main_agent=False,
+    ):
         super(SpawnAgentCommand, self).__init__()
         self._command_type = "SpawnAgent"
         self.set_location(location)
         self.set_rotation(rotation)
         self.set_type(agent_type)
         self.set_name(name)
+        self.add_number_parameters(max_height)
         self.add_number_parameters(int(is_main_agent))
 
     def set_location(self, location):